fix: moving 'midialinks' folder to footer folder
